requirements.txt
streamlit
cmake
opencv-python-headless
numpy
dlib-bin
imutils

